Clean up dax dialog tests and lingering fire button education code
